added 404 tests to PanelAPITest

+++ caliber/src/main/java/com/revature/caliber/controllers/PanelController.java
@@ -21,6 +21,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.revature.caliber.beans.Panel;
 import com.revature.caliber.services.PanelService;
+import com.revature.caliber.services.TrainingService;
 
 /**
  * @author Connor Monson
@@ -31,6 +32,8 @@ public class PanelController {
 
 	private static final Logger log = Logger.getLogger(PanelController.class);
 
+	@Autowired
+	private TrainingService trainingService;
 	@Autowired
 	private PanelService panelService;
 
@@ -44,7 +47,7 @@ public class PanelController {
 	public ResponseEntity<List<Panel>> findAll() {
 		log.debug("Getting all panels");
 		List<Panel> panels = panelService.findAllPanels();
-		HttpStatus status = (panels.isEmpty()) ? HttpStatus.NO_CONTENT : HttpStatus.OK;
+		HttpStatus status = panels.isEmpty() ? HttpStatus.NO_CONTENT : HttpStatus.OK;
 		return new ResponseEntity<>(panels, status);
 	}
 
@@ -54,9 +57,10 @@ public class PanelController {
 	public ResponseEntity<Panel> findPanelById(@PathVariable int panelId) {
 		log.debug("Getting category: " + panelId);
 		Panel panel = panelService.findById(panelId);
-		HttpStatus status = (panel == null) ? HttpStatus.NO_CONTENT : HttpStatus.OK;
-		log.info(panel);
-		return new ResponseEntity<>(panel, status);
+		if (panel == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<>(panel, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/panel/trainee/{traineeId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -64,8 +68,11 @@ public class PanelController {
 	@PreAuthorize("hasAnyRole('VP', 'QC', 'TRAINER', 'STAGING')")
 	public ResponseEntity<List<Panel>> findPanelByTraineeId(@PathVariable int traineeId) {
 		log.debug("Getting category: " + traineeId);
+		if (trainingService.findTrainee(traineeId) == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 		List<Panel> panels = panelService.findByTraineeId(traineeId);
-		HttpStatus status = (panels.isEmpty()) ? HttpStatus.NO_CONTENT : HttpStatus.OK;
+		HttpStatus status = panels.isEmpty() ? HttpStatus.NO_CONTENT : HttpStatus.OK;
 		log.info(panels);
 		return new ResponseEntity<>(panels, status);
 	}
@@ -76,7 +83,7 @@ public class PanelController {
 	public ResponseEntity<List<Panel>> findAllRepanel() {
 		log.debug("Getting all panels with repanel");
 		List<Panel> panels = panelService.findAllRepanel();
-		HttpStatus status = (panels.isEmpty()) ? HttpStatus.NO_CONTENT : HttpStatus.OK;
+		HttpStatus status = panels.isEmpty() ? HttpStatus.NO_CONTENT : HttpStatus.OK;
 		return new ResponseEntity<>(panels, status);
 	}
 
@@ -92,7 +99,7 @@ public class PanelController {
 	@Transactional(isolation = Isolation.READ_COMMITTED, propagation = Propagation.REQUIRED)
 	@PreAuthorize("hasRole('VP')")
 	public ResponseEntity<Panel> savePanel(@Valid @RequestBody Panel panel) {
-		panelService.createPanel((panel));
+		panelService.createPanel(panel);
 		return new ResponseEntity<>(panel, HttpStatus.CREATED);
 	}
 
@@ -101,8 +108,10 @@ public class PanelController {
 	@PreAuthorize("hasAnyRole('VP', 'TRAINER')")
 	public ResponseEntity<Void> deleteAssessment(@PathVariable int panelId) {
 		log.info("Deleting panel: " + panelId);
+		if (panelService.findById(panelId) == null) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
 		panelService.deletePanel(panelId);
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
-
 }
